+++ org/antlr/v4/runtime/BufferedTokenStream.cpp
@@ -122,7 +122,7 @@ namespace org {
                 }
 
                 Token *BufferedTokenStream::get(int i) {
-                    if (i < 0 || i >= tokens.size()) {
+                    if (i < 0 || i >= (int)tokens.size()) {
                         throw IndexOutOfBoundsException(std::wstring(L"token index ") +
                                                         std::to_wstring(i) +
                                                         std::wstring(L" out of range 0..") +
@@ -137,7 +137,7 @@ namespace org {
                     }
                     lazyInit();
                     std::vector<Token*> subset = std::vector<Token*>();
-                    if (stop >= tokens.size()) {
+                    if (stop >= (int)tokens.size()) {
                         stop = (int)tokens.size() - 1;
                     }
                     for (int i = start; i <= stop; i++) {
@@ -172,7 +172,7 @@ namespace org {
 
                     int i = p + k - 1;
                     sync(i);
-                    if (i >= tokens.size()) { // return EOF token
+                    if (i >= (int)tokens.size()) { // return EOF token
                         // EOF must be last token
                         return tokens[tokens.size() - 1];
                     }
@@ -211,7 +211,7 @@ namespace org {
 
                 std::vector<Token*> BufferedTokenStream::getTokens(int start, int stop, std::vector<int> *types) {
                     lazyInit();
-                    if (start < 0 || stop >= tokens.size() || stop < 0 || start >= tokens.size()) {
+					if (start < 0 || stop >= (int)tokens.size() || stop < 0 || (int)start >= tokens.size()) {
                         throw new IndexOutOfBoundsException(std::wstring(L"start ") +
                                                             std::to_wstring(start) +
                                                             std::wstring(L" or stop ") +
@@ -251,7 +251,7 @@ namespace org {
                 int BufferedTokenStream::nextTokenOnChannel(int i, int channel) {
                     sync(i);
                     Token *token = tokens[i];
-                    if (i >= size()) {
+					if (i >= (int)size()) {
                         return -1;
                     }
                     while (token->getChannel() != channel) {
@@ -274,7 +274,7 @@ namespace org {
 
                 std::vector<Token*> BufferedTokenStream::getHiddenTokensToRight(int tokenIndex, int channel) {
                     lazyInit();
-                    if (tokenIndex < 0 || tokenIndex >= tokens.size()) {
+					if (tokenIndex < 0 || tokenIndex >= (int)tokens.size()) {
                         throw new IndexOutOfBoundsException(std::to_wstring(tokenIndex) +
                                                             std::wstring(L" not in 0..") +
                                                             std::to_wstring(tokens.size() - 1));
@@ -299,7 +299,7 @@ namespace org {
 
                 std::vector<Token*> BufferedTokenStream::getHiddenTokensToLeft(int tokenIndex, int channel) {
                     lazyInit();
-                    if (tokenIndex < 0 || tokenIndex >= tokens.size()) {
+					if (tokenIndex < 0 || tokenIndex >= (int)tokens.size()) {
                         throw new IndexOutOfBoundsException(std::to_wstring(tokenIndex) +
                                                             std::wstring(L" not in 0..") +
                                                             std::to_wstring(tokens.size() - 1));
@@ -361,7 +361,7 @@ namespace org {
                         return L"";
                     }
                     lazyInit();
-                    if (stop >= tokens.size()) {
+					if (stop >= (int)tokens.size()) {
                         stop = (int)tokens.size() - 1;
                     }
